Break apart template root module
Break lib.rs into `value` and `tests` submodules
